requirements.txt
python-telegram-bot==20.8
Flask
yt-dlp
aiohttp